Set default lookBackDate for testing when one isn't provided. (#2819)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4967

## Note
This is a copy of https://github.com/CMSgov/dpc-app/pull/2818. That
branch failed its SonarQube check and is being left as is so the Cloud
team can investigate. This new branch was created so we can merge
without having to rerun the failing test on the old one.

## 🛠 Changes

Set default `lookBackMonths` in `OperationsConfig` when one isn't
provided. If we don't provide a `lookBackMonths` it defaults to 0, and
that factored in with the changes introduced in
https://github.com/CMSgov/dpc-app/pull/2815 cause tests to fail when run
during the first few days of the month.

This is because the `MockBlueButtonClient` sets its `TEST_LAST_UPDATED`
date to three days ago, and a transaction from three days ago is from
last month during the first few days of the current month. If we have a
0 month look back check these results get filtered.

## ℹ️ Context

Tests are failing during the first three days of the month.

## 🧪 Validation

It's 11/3 and tests are passing.