Me/dpc 4943 connection pool config (#2821)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4943

## 🛠 Changes

- Makes connection pool settings on FHIR clients configurable.
- Sets default values for `poolMaxTotal` and `poolMaxPerRoute` to 20.
This is the default value provided by HAPI if we don't provide one.
- Updated BBClient config to use HAPI interface instead of going around
it direct to the Apache client.

## ℹ️ Context

We want to be able to configure connection pool settings to fine tune
multi-threaded performance. Two points to make:

- The default value provided by HAPI FHIR without us doing any config is
20, so there is no change to existing functionality or performance.
- If you want to change the value on a per environment basis (as
mentioned in the AC of the Jira ticket) you only have to set
`POOL_MAX_PER_ROUTE` and `POOL_MAX_TOTAL` in the environment's .env
file.

## 🧪 Validation

All tests passing, nothing changed.