DPC-4971: Port authentication smoke tests to K6 (#2816)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4971

## 🛠 Changes

- Adds k6 script that tests api authentication endpoints
- Adds k6 script to generate keys, snippets and jwts
- Adds script to test new generation script
- Adds apis to dpc-api-client.js
- Updates dpc-api-client.js for use in all environments
- New GHA to run k6 smoke tests
- Update dpc-load-test GHA to set environment
- Updates GHA load testing unit tests to make a little simpler
- Adds make commands for testing the jwt script and running k6 smoke
tests
- Updates elint config to accommodate 'crypto', which is available in k6
without import

## ℹ️ Context

We want to move our smoke test to K6. This is to move the auth parts.

k6 does not provide a jwt builder, and we can't use node, so we need to
make a custom builder.

Auth parts gleaned from where smoke tests register organizations and
upload the public key.
- calls APIAuthHelpers.generateAndUploadKey
[SmokeTest::generateAndUploadKey](https://github.com/CMSgov/dpc-app/blob/5dc6386a5c1403737687fbc79ce88916d22600c0/dpc-smoketest/src/main/java/gov/cms/dpc/testing/smoketests/SmokeTest.java#L236)
- calls FHIRHelpers.registerOrganization
[SmokeTest::registerOrg](https://github.com/CMSgov/dpc-app/blob/5dc6386a5c1403737687fbc79ce88916d22600c0/dpc-smoketest/src/main/java/gov/cms/dpc/testing/smoketests/SmokeTest.java#L256)

The smoke tests currently use jwts/access tokens for the entire flow,
but this seems unnecessarily fraught. Instead, this code builds a jwt,
uses it to get an access token, and uses the access token to make a call
to get organization info. This should validate the auth flow, while
still allowing the use of the admin macaroon for calls to the api.

## 🧪 Validation

Passes locally:
- make smoke/k6
- make start-jwt-tests
- make start-load-tests
- docker run --rm -v ./dpc-load-testing:/src -e ENVIRONMENT=local -i
grafana/k6 run /src/ci-app.js
(this last checks that ci-app.js still works)

And on GHA:
- Test:https://github.com/CMSgov/dpc-app/actions/runs/19105168012
- Sandbox: https://github.com/CMSgov/dpc-app/actions/runs/19105279333
- Unit Test: https://github.com/CMSgov/dpc-app/actions/runs/19048799413
- Load Test: https://github.com/CMSgov/dpc-app/actions/runs/19048844538

Post-refactor GHA test:

https://github.com/CMSgov/dpc-app/actions/runs/19639318493