Ja/dpc 4995 move testing utility classes to commons (#2839)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4995

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This change moves the utility class DBUtils.java in dpc-testing to the
dpc-commons module. This class is referenced in dpc-aggregation and
dpc-attribution source code. This change ensures that test only
libraries are not leaked into the runtime classpath of downstream
modules, and the dpc-testing module is treated as a pure test harness
module. Additionally, Its internal dependencies such as
dropwizard-testing, mockito-core, and other test framework libraries are
marked with `<scope>provided</scope>` in dpc-testing/pom.xml. This
allows the module to compile its own test utilities while preventing
these dependencies from being pulled transitively into the runtime of
services that consume dpc-testing.

## ℹ️ Context

This is part of the larger effort to identify and scope test
dependencies with the test scope.
[DPC-4946](https://jira.cms.gov/browse/DPC-4946)
<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

- Ensured that the unit test pass correctly and that the application
runs as expected locally.
- Confirmed the test dependency removals using artifact (JAR) Inspection
and Dependency Tree Analysis:
- Example Command: unzip -l dpc-attribution/target/dpc-attribution.jar |
grep -i mockito
- Example Command: mvn dependency:tree -Dscope=runtime | grep -i mockito
- Confirmed GitHub Actions worfklow passes against dev:
https://github.com/CMSgov/dpc-app/actions/runs/19644029612

NOTE on Dependency Scope Exception: The dependency
org.opentest4j:opentest4j:compile is still appearing on the runtime
classpath. This is intentional and necessary because it is a transitive
dependency of org.junit.jupiter, which is set to <scope>compile</scope>
in the dpc-smoketest POM. This is left as is to prevent build issues in
our GitHub Actions workflow, which currently relies on building a
standalone executable JAR for the Java smoke tests. This issue will be
fully resolved and the dependency removed once the Java smoke tests are
deprecated and replaced by the k6 smoke test suite.

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->