DPC-5002 consolidate k6 smoke tests (#2843)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5002

## 🛠 Changes
- remove k6 portals test from jmeter smoke tests
- instructions updated
- add portals smoke test workflow to k6 smoke test
- configure portals to accept calls as host.docker.internal
- configure smoke/k6 to use existing golden macaroon, if present

## ℹ️ Context

We are moving our smoke tests from jmeter to k6. The test have been
written separately; now they are consolidated.

I switched back to using 3 vus instead of 3 iterations, because the
tests should run faster and in parallel with vus, and as long as we keep
3, it should distribute evenly.

The portals test now just check that calling a protected path redirects
to the sign in page, as k6 is easier to configure that way than jmeter,
and that is what it really tests. I switched from root to
organizations/foo/edit in the web portal (aka sandbox user portal)
because root does not redirect, and I wanted to keep the tests the same
for simplicity.

## 🧪 Validation

- pass locally (please check)
- pass on dev:
https://github.com/CMSgov/dpc-app/actions/runs/19837967505
- pass on test:
https://github.com/CMSgov/dpc-app/actions/runs/19837976891
- pass on sandbox:
https://github.com/CMSgov/dpc-app/actions/runs/19837977916
- pass on prod:
https://github.com/CMSgov/dpc-app/actions/runs/19837982739

---------

Co-authored-by: MEspositoE14s <133133846+MEspositoE14s@users.noreply.github.com>