[DPC-5041] Log exceptions in dpc-api (#2854)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5041

## 🛠 Changes

Ensures that exceptions are logged to dpc-api and successful health,
version, and ping checks are not.

## ℹ️ Context

A previous ticket removed healthcheck from our logs, for ease of use,
but a Dropwizard logging filter was also removing all exception traces.
This PR resolves the exception issue, as well as adds `/ping` to the
list of successful checks that need not be logged.

## 🧪 Validation

Tests updated, plus visual inspection of the smoke test logs.