DPC-4858 JS instead of Browser validation of CD invite form (#2852)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4858

## 🛠 Changes

- CD invitation form uses JS validation instead of browser validations
- text input component has id field in error message to support JS
validation
- Specs updated for necessary changes to text input component

## ℹ️ Context

We want a consistent look and feel in the validation of this form.
Browser validations worked for all but equivalent value of email and
email confirmation. At the current state of my knowledge, Javascript
does not have access to browser validation UI, so we emulated
Rails-style error UI. To avoid this inconsistency, we are now using
Javascript to ensure each error message looks like the others.

## 🧪 Validation

Manual testing:


https://github.com/user-attachments/assets/4526d6d5-f8f5-4236-bd77-d25464b90d3c