Me/dpc 5022 fix http responses (#2858)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5022

## 🛠 Changes

When attempting a request that uses a path authorizer, we no longer
differentiate between not authorized (401) and not found (404). In both
cases we return a 404 and a description explaining what happened.

## ℹ️ Context

After uploading and then deleting a practitioner, we noticed that trying
to delete it again returned "not authorized" instead of "not found". For
security reasons, we don't want to tell the caller whether a resource
exists if they're not authorized to see it, so in both cases we respond
with a "not found".

## 🧪 Validation

- Ran locally and verified results.
- Added IT that validates response.

### Note:
If you test this, you have to use a real UUID. Our endpoints are defined
as `resource/UUID`, so if you send
`resource/some_string_thats_not_a_uuid` you'll get a generic 404 with no
operation outcome since you're asking for an endpoint that doesn't
exist. It may be worth changing this in the future.