DPC-5053 Fix no account login flow (#2862)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5053

## 🛠 Changes

- Invitation error component renamed so can be used more generally
- Login IAL/1 without existing user shows error message with ability to
log out of login.gov

## ℹ️ Context

If a user without an account tries to log in to the portal from the sign
in page (rather than via the invitation flow), we reject the attempt.
However, as the user was still logged in to login.gov, hitting the log
in button just sent them to login.gov, which redirected them back to us,
which led them back to the sign in page. Now, instead of the sign in
page, we send them to an error page with the option to log out of
login.gov.

As we already have a fairly generic error page set up for the invitation
flow, I just used that.

## 🧪 Validation

Manual and automated testing.
### No account
<img width="846" height="456" alt="no_account"
src="https://github.com/user-attachments/assets/1ea7a62e-29f1-4edd-b512-bab9b28820a2"
/>

### Failed Invite
<img width="834" height="419" alt="failed_invite"
src="https://github.com/user-attachments/assets/8efaafd6-84f9-49e4-9679-8cd82f61c804"
/>