DPC-5103 fix check healthy (#2866)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5103

## 🛠 Changes

- added skip value to matrix in check_deploy
- added skip possibility in check-deployment-status script

## ℹ️ Context

Changes to monitoring ecs deploy included removing a not-great mechanism
for skipping the check for services that are not deployed to a given
environment -- for example, we don't deploy dpc-admin or dpc-web to
prod. This PR makes deciding which services to check in deployment
explicit in the check_deploy matrix.

## 🧪 Validation

Ran truncated versions of ecs-deploy to verify things run unless skipped
(used "test" instead of "sandbox" for actual test).

- All checks run:
https://github.com/CMSgov/dpc-app/actions/runs/20467399219/job/58814423116
- Portal skipped:
https://github.com/CMSgov/dpc-app/actions/runs/20467410691/job/58814458362