[DPC-5105] Upgrade performance of unique_providers_served_all_time view (#2868)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-5105

## 🛠 Changes

Upgrade `unique_providers_served_all_time` view for better performance.

## ℹ️ Context

The `unique_providers_served_all_time` view runs in a few seconds in
dev, test and sandbox, but hits the 20 minute time out in prod. This new
version runs in prod in about 2 seconds.

## 🧪 Validation

- Ran the underlying query in prod to test performance.
- Ran the query against our persisted environments to verify the results
are the same as the existing view.
- Successfully deployed to dev
[here](https://github.com/CMSgov/dpc-app/actions/runs/20489128182).