Me/dpc 5062 table org formatting (#2870)
## 🎫 Ticket

[DPC-5062](https://jira.cms.gov/browse/DPC-5062) 

## 🛠 Changes

Adds new css class `header-top-border-in-body` that adds the top border
back to `<th>` elements within a table body and uses it on the org list.

## ℹ️ Context

At some point, USWDS removed the top border from `<th>` elements in
`usa-table--borderless`, which makes sense on its face since the table's
supposed to be borderless. The problem is that we use a `<th>` as the
first column in our org table, and removing the top border messes up the
line between the top row and the header. This change adds the top border
back in so we get a uniform line across the table.

An easier solution would've been to change our `<th>` to a `<td>`, but
then SonarQube complains that `<td>`s can't have the `scope` attribute
and we need that to be compatible with screen readers.

## 🧪 Validation

- Deployed to
[dev](https://github.com/CMSgov/dpc-app/actions/runs/20601441198)
- Tested locally.

Before:
<img width="899" height="394" alt="before"
src="https://github.com/user-attachments/assets/fc3581f7-7813-4827-bd22-19198de08b55"
/>


After:
<img width="975" height="439" alt="after"
src="https://github.com/user-attachments/assets/31fd6973-766d-49fc-a5d8-ab94b94c1ad3"
/>